+++ frontend/fingertips-frontend/components/organisms/LineChartTable2/index.tsx
